search_by_gene: adding a limit of 10 proteins per species, with a show more button
